+++ src/main/java/com/woongeya/zoing/domain/like/domain/repository/LikeRepository.java
@@ -10,6 +10,6 @@ public interface LikeRepository extends JpaRepository<Like, Long> {
     Optional<Like> findByUserIdAndProjectId(Long userId, Long projectId);
     List<Like> findByProjectId(Long projectId);
     List<Like> findByUserId(Long userId);
-    Long countByUserId(Long userId);
-    Long countByProjectId(Long projectId);
+    Integer countByUserId(Long userId);
+    Integer countByProjectId(Long projectId);
 }
